+++ vehicle/src/Vehicle/Data/Builtin/Loss/Instances.hs
@@ -291,7 +291,7 @@ ifCandidates =
 existsCandidates :: [TempCandidate mode]
 existsCandidates =
   [ ( hasExists tRatWithGradients,
-      builtinFunction SearchRatTensor,
+      builtinFunction (SearchRatTensor Exists),
       Nothing
     ),
     ( hasExists tBool,